added unit tests for compose and modularized context into contexts
