+++ lib/compose.js
@@ -1,5 +1,7 @@
-const {UnaryContext, ClientStreamContext, ServerStreamContext, DuplexContext} =
-    require('./context/context');
+const UnaryContext = require('./contexts/unaryContext')
+const ClientStreamContext = require('./contexts/clientStreamContext')
+const ServerStreamContext = require('./contexts/serverStreamContext')
+const DuplexContext = require('./contexts/duplexContext');
 
 const getMethodType = require('./utils/getMethodType');
